move credential test
